requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql
python-jose
passlib[bcrypt]
email-validator
alembic
mako
pydantic-settings
python-multipart
httpx
pytest
requests
gunicorn
cryptography
